requirements.txt
pypdfium2==4.30.0
python-docx==1.1.2
paddleocr==2.7.0.3
paddlepaddle==2.6.2
opencv-python==4.6.0.66
numpy==1.26.4
pillow==10.4.0
shapely==2.0.4
regex==2024.5.15
